Splits useProducts from ProductContext for Fast Refresh functionality
