+++ src/components/NavBar/index.jsx
@@ -1,7 +1,7 @@
 // NavBar/index.jsx
 import PropTypes from "prop-types";
 import CartIcon from "../CartIcon";
-import { useProducts } from "../../contexts/ProductContext";
+import { useProducts } from "../../contexts/ProductContext/useProducts";
 import * as S from "./index.styled"; // Import styled components with alias S
 
 function NavBar({ itemCount }) {
