+++ src/contexts/ProductContext/useProducts.js
@@ -1,0 +1,5 @@
+// useProducts.js
+import { useContext } from "react";
+import { ProductContext } from "../ProductContext";
+
+export const useProducts = () => useContext(ProductContext);
